### 2019-09-02 #45 @whedon
[PRE REVIEW]: Fashionable Modelling With Flux
**Submitting author:** @dhairyagandhi96 (<a href="http://orcid.org/0000-0002-8549-5270">Dhairya Gandhi</a>)
**Repository:** <a href="https://github.com/dhairyagandhi96/flux-juliacon-paper.git" target ="_blank">https://github.com/dhairyagandhi96/flux-juliacon-paper.git</a>

Thanks for submitting your paper to JuliaCon @dhairyagandhi96. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@dhairyagandhi96 if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
