### 2019-10-17 #56 @whedon
[REVIEW]: Fashionable Modelling With Flux
**Submitting author:** @dhairyagandhi96 (<a href="http://orcid.org/0000-0002-8549-5270">Dhairya Gandhi</a>)
**Repository:** <a href="https://github.com/dhairyagandhi96/flux-juliacon-paper.git" target ="_blank">https://github.com/dhairyagandhi96/flux-juliacon-paper.git</a>
