### 2019-11-01 #56 @denizyuret
Some questions:
- [ ] I am not sure what the page limit is. Assuming 8 pages is ok.
- [ ] The repository url points to the paper repo, not Flux.jl. Assuming Flux.jl for other questions.
- [ ] The formatting seems off, lots of space at the bottom of each page, is this normal?
