### 2019-11-05 #56 @dhairyagandhi96
Thanks for the review. Following this, we've made some changes to address some of the suggestions.

Re the AlphaGo point; we couldn't converge it for Go.

Per section 3; the GPU functionality, as well as the AD are mature. The acceleration with TPUs, outputting with JS and batching are working prototypes as of now. We plan to expand this functionality in the future.

With regards to Listing 7, it is not meant as a motivating example, rather an illustrative one. Broadcasting would do the same work internally, the vectorisation of the code is the larger question that we've tried to address in that section.

The missing cite to Knet is unintentional and we shall remedy that.
